requirements.txt
python-nmap 
vulners
tabulate
